Add StockAnalysisView components and resolve conflicts in CSRF and StockAnalysisAssistant
