+++ src/components/CSRFToken/CSRFToken.js
@@ -15,3 +15,4 @@ const CSRFToken = () => {
 
 export default CSRFToken;
 
+
