+++ src/components/StockAnalysisView/ConclusionSection/ConclusionSection.js
@@ -1,0 +1,6 @@
+// const ConclusionSection = ({ summary }) => (
+//     <div>
+//         <h3 className="text-lg font-bold">📝 Summary</h3>
+//         <p className="text-gray-300 mb-2">{summary || "No data"}</p>
+//     </div>
+// );
