+++ src/components/StockAnalysisView/DetailBox/DetailBox.js
@@ -1,0 +1,6 @@
+// const DetailBox = ({ label, content }) => (
+//     <div className="flex flex-col">
+//         <span className="block font-bold">{label}</span>
+//         <span className="text-gray-300">{content || "No data"}</span>
+//     </div>
+// );
