+++ src/components/StockAnalysisView/DetailSection/DetailSection.js
@@ -1,0 +1,9 @@
+
+
+
+// const DetailSection = ({ title, content }) => (
+//     <div className="border-t border-b border-custom-purple py-4 flex flex-col">
+//         <h3 className="text-lg font-bold">{title}</h3>
+//         <p className="text-gray-300">{content}</p>
+//     </div>
+// );
