+++ src/components/StockAnalysisView/EMABox/EMABox.js
@@ -1,0 +1,6 @@
+// const EMABox = ({ label, value }) => (
+//     <div className="p-3 bg-custom-purple rounded-lg flex flex-col">
+//         <span className="font-semibold">{label}</span>
+//         <span>{value || "No available data"}</span>
+//     </div>
+// );
