+++ src/components/StockAnalysisView/FinancialsSection/FinancialSection.js
@@ -1,0 +1,10 @@
+// const FinancialsSection = ({ data }) => (
+//     <div className="border-b border-custom-purple pb-4">
+//         <h3 className="text-lg font-bold">💰 Financials</h3>
+//         <div className="grid grid-cols-3 gap-4">
+//             <DetailBox label="📈 Revenue:" content={data?.quarterly_revenue} />
+//             <DetailBox label="💵 Net Income:" content={data?.quarterly_net_income} />
+//             <DetailBox label="📊 EPS:" content={data?.quarterly_eps} />
+//         </div>
+//     </div>
+// );
