+++ src/components/StockAnalysisView/RecommendationsSection/RecommendationSection.js
@@ -1,0 +1,9 @@
+// const RecommendationsSection = ({ data }) => (
+//     <div className="border-b border-custom-purple pb-4">
+//         <h3 className="text-lg font-bold">🎯 Recommendations</h3>
+//         <div className="space-y-2">
+//             <DetailBox label="💭 Considerations:" content={data?.considerations} />
+//             <DetailBox label="⚡ Action:" content={data?.recommendations_for_action} />
+//         </div>
+//     </div>
+// );
